requirements.txt
gunicorn==19.3.0
Django==1.8.3
PyYAML==3.11
django-debug-toolbar==1.3.2
dj-database-url==0.3.0
dj-static==0.0.6
psycopg2==2.6.1
django-nopassword==2.1.0
django-reversion==1.9.3
coverage==3.7.1
python-dateutil==2.4.2